move HazardousWasteForm to directory
