+++ client/src/components/Manifest/WasteLine/HazardousWasteForm/index.ts
@@ -1,0 +1,3 @@
+import { HazardousWasteForm } from 'components/Manifest/WasteLine/HazardousWasteForm/HazardousWasteForm';
+
+export { HazardousWasteForm };
